Extract decodeMsg into separate file
