+++ packages/iov-bns/src/context.ts
@@ -10,7 +10,7 @@ import {
   SwapProcessState,
 } from "@iov/bcp";
 
-import { asIntegerNumber, decodeAmount, ensure } from "./decode";
+import { asIntegerNumber, decodeAmount, ensure } from "./decodinghelpers";
 import * as codecImpl from "./generated/codecimpl";
 import { Keyed } from "./types";
 import { addressPrefix, encodeBnsAddress } from "./util";
