+++ packages/iov-bns/src/decodinghelpers.spec.ts
@@ -1,0 +1,80 @@
+import { TokenTicker } from "@iov/bcp";
+
+import { decodeAmount } from "./decodinghelpers";
+import * as codecImpl from "./generated/codecimpl";
+import { coinBin, coinJson } from "./testdata.spec";
+
+describe("decodinghelpers", () => {
+  describe("decodeAmount", () => {
+    it("can decode amount 3.123456789 ASH", () => {
+      const backendAmount: codecImpl.coin.ICoin = {
+        whole: 3,
+        fractional: 123456789,
+        ticker: "ASH",
+      };
+      const decoded = decodeAmount(backendAmount);
+      expect(decoded).toEqual({
+        quantity: "3123456789",
+        fractionalDigits: 9,
+        tokenTicker: "ASH" as TokenTicker,
+      });
+    });
+
+    it("can decode amount 0.000000001 ASH", () => {
+      const backendAmount: codecImpl.coin.ICoin = {
+        whole: 0,
+        fractional: 1,
+        ticker: "ASH",
+      };
+      const decoded = decodeAmount(backendAmount);
+      expect(decoded).toEqual({
+        quantity: "1",
+        fractionalDigits: 9,
+        tokenTicker: "ASH" as TokenTicker,
+      });
+    });
+
+    it("can decode max amount 999999999999999.999999999 ASH", () => {
+      // https://github.com/iov-one/weave/blob/v0.9.3/x/codec.proto#L15
+      const backendAmount: codecImpl.coin.ICoin = {
+        whole: 10 ** 15 - 1,
+        fractional: 10 ** 9 - 1,
+        ticker: "ASH",
+      };
+      const decoded = decodeAmount(backendAmount);
+      expect(decoded).toEqual({
+        quantity: "999999999999999999999999",
+        fractionalDigits: 9,
+        tokenTicker: "ASH" as TokenTicker,
+      });
+    });
+
+    it("is compatible to test data", () => {
+      const decoded = codecImpl.coin.Coin.decode(coinBin);
+      const amount = decodeAmount(decoded);
+      expect(amount).toEqual(coinJson);
+    });
+
+    it("throws for decoding negative amount", () => {
+      // ICoin allows negative values, which are now supported client-side
+      {
+        // -3.0 ASH
+        const backendAmount: codecImpl.coin.ICoin = {
+          whole: -3,
+          fractional: 0,
+          ticker: "ASH",
+        };
+        expect(() => decodeAmount(backendAmount)).toThrowError(/`whole` must not be negative/i);
+      }
+      {
+        // -0.123456789 ASH
+        const backendAmount: codecImpl.coin.ICoin = {
+          whole: 0,
+          fractional: -123456789,
+          ticker: "ASH",
+        };
+        expect(() => decodeAmount(backendAmount)).toThrowError(/`fractional` must not be negative/i);
+      }
+    });
+  });
+});
